Move utility functions to role.js and add functions.logger.log() to autoAdminOnRegister
