+++ functions/src/utils/role.js
@@ -1,5 +1,19 @@
 const admin = require("firebase-admin");
 
+/** Remove a user from the list of admins
+ *
+ * @param {string} adminId
+ */
+const removeAdmin = async (adminId) => {
+  await admin
+    .firestore()
+    .collection("roles")
+    .doc("admin")
+    .update({
+      users: admin.firestore.FieldValue.arrayRemove(adminId),
+    });
+};
+
 /** Add a user to the list of admins
  *
  * @param {string} newAdminId
@@ -15,3 +29,4 @@ const addAdmin = async (newAdminId) => {
 };
 
 exports.addAdmin = addAdmin;
+exports.removeAdmin = removeAdmin;
